Handle fields which are unique with sort-order changing
Changing only a field which is unique with sort_order needs to reset sort order (or you could break the uniqueness condition).

I later realised this fix is similar to what @mjtamlyn did in #26, but is narrower in focus, and maintains 1.4 compatibility.

Therefore I'm committing to a branch so I can use temporarily for a 1.4 project until we upgrade.